requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.4
django-ckeditor==6.4.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
gunicorn==20.1.0
numpy==1.22.4
opencv-python==4.5.5.64
Pillow==9.1.1
psycopg2==2.9.3
python-decouple==3.6
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.1.0
